### 2020-07-18 #2493 @whedon
[PRE REVIEW]: CIMSA: Customizable Integrated Multiple Sequence Alignment Package
**Submitting author:** @kuangmeng (<a href="http://orcid.org/0000-0001-9656-5652">Mengmeng Kuang</a>)
**Repository:** <a href="https://github.com/kuangmeng/CIMSA" target ="_blank">https://github.com/kuangmeng/CIMSA</a>
**Version:** v1.0
