### 2020-07-18 #2493 @labarba
👋  @kuangmeng — I'm going to send this back for a bit of work on the paper. Currently, it does not follow our guidance in providing a clear _Statement of Need_. See the section [What should my paper contain?](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) in our documentation. Tell us why the research world needs your software!

Also, to be eligible for consideration in JOSS, your software must be under a standard public license, approved by the Open Source Initiative. See our [documentation](https://joss.readthedocs.io/en/latest/review_criteria.html#software-license).
