### 2020-07-28 #2493 @will-rowe
Hi @kuangmeng

Thanks for the submission! I've been reading through it this morning and have a couple of things I'd like before we put this out for review.

* I'm not an MSA algorithm researcher so am finding this a little hard to find a use case for. It would be great if you could provide a usage example, ideally for a real-world analysis problem. This will also help me find reviewers for your submission. This would be really beneficial at the top of the README.

* Please could you add some additional testing. I see that you have a test to make sure the compiled binary works with all available parameters, this is good! But I think we really need some unit testing - particularly where you have reimplemented MSA algorithms. It would be great if you could also hook this up to continuous integration (e.g. travis)

* I think the documentation needs a little work beyond what you have in the README. It would be really useful to have some more information on the various parameters, what the methods you have reimplemented are doing and how to access/interpret the results. Your figure from the paper would be useful here too. To save the README getting too crowded, you could add documentation via GitHub wiki or a service like readthedocs.

Sorry that these are quite significant points, but I really think they need doing before we can review this. Let me know if you need any further clarification or help.
