### 2020-08-01 #2493 @kuangmeng
Hi @will-rowe

Thanks for your comments and I have revised the README file for your 1st and 3rd comments. As for 2nd comment, I will try my best to do it.
