### 2020-09-03 #2493 @kuangmeng
Dear @will-rowe ,

Thank you for waiting so long. Since the calculation results of each step of my CIMSA are uncertain, I can't add normal unit tests. Therefore, in order to determine the correctness of each step, I add the `assert()` statement and print the intermediate running state. I hope this can help you understand my program better. Thanks.
