### 2020-09-03 #2493 @will-rowe
Hi @kuangmeng - thanks for your update. My concern is that this still falls short of our reviewer [criteria](https://joss.readthedocs.io/en/latest/review_criteria.html?highlight=tests#tests):

```
No way for you, the reviewer, to objectively assess whether the software works
```

This might be my fault for not being specific enough. Could you include some sort of "truth set". For example, a set of sequences and the expected MSA? These expected MSAs could be generated using other tools (maybe clustal for instance?). A minimal example could be to have a test that takes an input set, produces an MSA using your software and then diffs this against a reference MSA.

[Here](https://github.com/HPCTUSofiaOrg/clustalw2/blob/master/test_clustal.sh) is an example I just found of another MSA aligner doing this. Personally, I'd prefer it done using a testing framework, and for fail/edge cases to be considered, but that is not a requirement for review (but might make your life easier when maintaining the software :))
