### 2020-09-14 #2493 @will-rowe
Hi @bblum9, @lmvasquezg and @emilydolson - would any of you be able to review this submission for JOSS please?
