### 2020-09-14 #2493 @xyzmjf
Hi Will

Apologies for taking a few days to reply. I can review this manuscript.
What timeline are you expecting, and what format would you want the review
in.
As this is the first time I will have reviewed for you, I do have a few
'newbie' questions.

Mark
