### 2020-09-14 #2493 @will-rowe
Perfect - thanks @xyzmjf

Please find the full reviewer documentation [here](https://joss.readthedocs.io/en/latest/review_criteria.html).

We will start a new thread in a moment which will record the submission process and this one will be closed. You can use the new thread to track the review process, mark off your reviewer checklist and let us know your recommendation etc. As well as this, we suggest writing your reviewer comments and feedback as a Github issue on the submitting author's repository (https://github.com/kuangmeng/CIMSA).

I'll start the review thread in a moment and please feel free to ask any question there. We are pretty flexible on how long the review takes; usually it's best if it can be received within a few weeks but I appreciate that timelines are tricky for everyone at the moment.
