### 2020-09-23 kuangmeng/CIMSA#1 @CFGrote
Build instructions
Documentation review (cf. https://github.com/openjournals/joss-reviews/issues/2672)

The `readme.md` gives the following instructions for building the code:
```
cd cimsa_folder/
make clean
make
```
For the inexperienced user, it may be unclear what `cimsa_folder` refers to, as there is obviously no directory of that name after downloading and extracting the zip archive or cloning the repo.

The second and third step of above return successfully.

The `readme.md` then continues with

"If you want to display the results using MSAViewer"
```
cd cimsa_folder/display_msa/
node start_display.js
```
It remains unclear if this step is a build step and needed to make certain parts of the package accessible (a viewer?) or if it is completely mandatory

The `node` command does not return anything, the process seems to hang. What should the user expect?
