### 2020-09-23 kuangmeng/CIMSA#2 @CFGrote
API documentation missing
Documentation review (cf. openjournals/joss-reviews#2672)

I cannot find a reference manual, neither for the c++ code, nor for python. I recommend doxygen for c++ and sphinx for python.
