### 2020-09-23 kuangmeng/CIMSA#3 @CFGrote
Automated testing missing
Documentation review (cf. openjournals/joss-reviews#2672)

There seems to be no automated testing in place. travi-ci.org would be one option.
