### 2020-09-23 kuangmeng/CIMSA#4 @CFGrote
Community guidelines missing
Documentation review (cf. openjournals/joss-reviews#2672)

There are no community guidelines in the readme.
