### 2020-09-23 kuangmeng/CIMSA#5 @CFGrote
Structure of `readme.md`
Documentation review (cf. openjournals/joss-reviews#2672)

The section "Test" should in my opinion proceed the section "Run" in `readme.md`. Clearly, a user would want to check the installation before starting her own runs.
