### 2020-09-23 kuangmeng/CIMSA#6 @CFGrote
Linked presentation slide
Documentation review (cf. openjournals/joss-reviews#2672)

The linked presentation is in my opinion not a good resource for more information on MSA. It stems from a multipart lecture series, refers to earlier parts of the series without direct links, does not contain any literature references, is full of specialist jargon ("gap penalties", "scoring function", "motifs", "dynamic programming") not suitable for non-experts and (typical for presentations) sparse on descriptive text. A reference to a wikipedia page on MSA and links therein seems more appropriate.
