### 2020-09-23 kuangmeng/CIMSA#7 @CFGrote
Command-line interface (CLI)
Documentation review (cf. openjournals/joss-reviews#2672)

The CLI accepts 5 "step" arguments with integer parameters. These integer parameters then map to certain algorithms or algorithmic variants, e.g. -s1 0 for "Double Affine Pair-HMMs" etc.  Here, I would recommend to use a more self-explanatory syntax, e.g. parameter names instead of integers. These names could be, e.g., "double_affine_hmm", "partition_function", "local_pair_hmm" etc for the -s1 argument, "min", "max", "avg", "length", "aligned_per_length" for -s2 and so forth. Certainly, this means more typing for the user, but the interface would become much clearer and errors more easy to spot. An example command line would then read
```
python CIMSA.py -s1 double_affine_hmm -s2 max -s3 upgma -s4 unweighted -s5 iterative
```
I further recommend to use the `argparse` python package (https://pypi.org/project/argparse/) to administer the command line arguments in `CIMSA.py`.
