### 2020-09-23 kuangmeng/CIMSA#8 @CFGrote
Testing
Documentation review (cf. openjournals/joss-reviews#2672)

In `test_cimsa.py`, the deprecated `os.system` is used. I recommend using the `subprocess` builtin module. Indeed, `subprocess` is imported but not used, just as a few other modules.

The function `MakeCMD` in `test_cimsa.py` reimplements the function of the same name in `CIMSA.py`. Hence, not the main python module is tested but only the reimplementation. I suggest adding  `from CIMSA import MakeCMD` and remove lines 24-26 in `test_cimsa.py`.

It would also be helpful to add a regression test, i.e. a test that takes some real-world sequence data, processes the data in CIMSA with a suitable combination of progressive MSA steps and then compares the output to a reference alignment of the same input data. In this way, developers will be able to check if their code changes introduced any bugs.
