### 2020-09-23 kuangmeng/CIMSA#9 @CFGrote
Statement of need
Paper review (cf. openjournals/joss-reviews#2672)

The author states that "Till now, there is no efficient and accurate MSA tool of progressive strategy using Python programming language or providing a Python interface." A web search on "multiple sequence alignment python" returns numerous relevant hits, therefore this statement seems at first glance unfounded. It only becomes understandable when the reader recognizes that the term "progressive" actually characterizes a particular family of MSA algorithms. In this light, the statement of need is appropriate but

- [ ] I recommend adding the term "progressive" already in the title and most definitely in the first sentence of the paper to avoid this sort of misunderstanding.
- [ ] An extended discussion and definition of "multiple sequence alignment" and its variant "progressive MSA" also seems in order.
