### 2020-09-23 kuangmeng/CIMSA#10 @CFGrote
State of the field
Paper review (cf. openjournals/joss-reviews#2672)

The paper mentions a number of other software packages that implement the progressive MSA algorithm. It compares the current package CIMSA to the extend  that the former are implemented in c++ and that CIMSA encapsulates a superset of algorithm variants which thereby become accessible through a common CLI. Given the purpose of CIMSA this may be sufficient.
However, a benchmark comparison of CIMSA to at least one other progressive MSA code using
 a given reference sequence dataset would tremendously increase the informative content of the paper and potentially demonstrate a) reliablity of CIMSA ("reproduces established codes' results") and b) performance ("finishes in shorter/same/longer" compute time).
