### 2020-09-23 kuangmeng/CIMSA#11 @CFGrote
Quality of writing
Paper review (cf. openjournals/joss-reviews#2672)

I strongly recommend to seek assistance from a english language native speaker or experienced academic writer.

Many remarks regarding python vs. c++ are made in repetitive fashion. The statement "[c++ code is] difficult to read and hard to understand [...]" seems too general.

What is meant by the term "eye-catching"? Does it mean "promising (to solve a certain problem)", or "beautiful" in some sense (to be defined)?
