### 2020-09-23 kuangmeng/CIMSA#12 @CFGrote
References
Paper review (cf. openjournals/joss-reviews#2672)

The list of references could be amended. I suggest adding references after
* first mention of MSA  (Summary)
* first mention of progressive MSA (Summary)
* "MSA, as a comparative basis in Bioinformatics, has received much attention in the past few decades" (Summary).
* first mention of HMMs (Design)

Furthermore, and to allow differentiating the author's own contributions from earlier developments, I would add references to
* double affine and local HMMs
* partition function
* UPGMA, WPGMA
* PCT
