### 2020-09-23 kuangmeng/CIMSA#13 @CFGrote
c++ vs python parts
General comments (openjournals/joss-reviews#2672)

CIMSA is a two part project: A c/c++ code to implement the progressive MSA algorithm and its variants and a python module `CIMSA.py` which acts as a frontend and exposes the command line interface. This separation into frontend and backend should be emphasized more clearly.

 Indeed, the python module aggregates a call to the compiled c++ code which has exactly the same command line syntax as the python module. It remains a bit opaque what the benefit of the extra python layer is. The benefit may be that a user can start a CIMSA run from within a python script or notebook, but this is not mentioned nor is such a usage scenario documented in the paper or readme.
