### 2020-09-23 kuangmeng/CIMSA#14 @CFGrote
Emphasize purpose of CIMSA more clearly
openjournals/joss-reviews#2672

The strength of CIMSA lies in the integration of numerous different variants of the progressive MSA algorithm and the harmonization of the user interface to all these variants. Indeed, one would have to spend countless hours to run all variants of progressive MSA on a given dataset even if one had all the established implementations installed and functional on a computer system. I feel that this benefit is not emphasized strongly enough in the paper.
