### 2020-09-23 kuangmeng/CIMSA#15 @CFGrote
Auhorship
General checks (cf. openjournals/joss-reviews#2672)

The submitting author has made major if not all contributions to the software. However,
the paper acknowledges contributions from other developers and researchers but these remain unnamed. Hence, the completeness question in openjournals/joss-reviews#2672 cannot be answered.
