### 2020-09-23 kuangmeng/CIMSA#16 @CFGrote
Statement of need in documentation
Documentation review (cf. openjournals/joss-reviews#2672)

An explicit statement of need is missing in the `readme.md` file. The text briefly explains the concept of MSA on a rather abstract bioinformatical level without reference to applications in e.g., genomics or proteomics.
