### 2020-09-23 kuangmeng/CIMSA#17 @CFGrote
List of dependencies
Documentation review (cf. openjournals/joss-reviews#2672)

A list of dependencies is absent. No automated package management solution. At a minimum, the cpp standard version (11) should be mentioned.
