### 2020-09-23 kuangmeng/CIMSA#18 @CFGrote
Example usage
Documentation review (openjournals/joss-reviews#2672)

A usage example is not defined as such. The `readme.md` states the general CLI syntax and the python test module `test_cimsa.py` iterates over all possible combinations of CLI arguments. I recommend to add a real-world example that solves an actual use case.
