### 2020-09-23 kuangmeng/CIMSA#19 @CFGrote
Recommendation to JOSS editors
The software package `CIMSA` provides a python and command line interface to a variety of
so-called "progressive multiple sequence alignment" algorithms, often used for genetic and protein sequence alignment in bioinformatics, genetics, and proteomics. It thereby provides a harmonized interface to a plethora of combinations for the various possible choices in the 5 steps of progressive MSA. Such an interface is to my knowledge new and deserves attention. I can see huge benefits in systematic exploration of parameter spaces and comparative benchmark studies in order to enhance the understanding of how to choose the best combination of step choices for a given MSA problem or dataset.

Nevertheless, in my opinion, the current state of the project does not warrant publication in JOSS. The software is not documented to a sufficient degree. The paper fails to make clear the need for such a software (which undoubtedly exists, see my comments above), certain key requirements for acceptance are not met.

Given the current state of the project `CIMSA`, my recommendation is to reject the submission and reconsider it after major revisions as indicated by issues #1 through #18.
