### 2020-09-23 #2672 @CFGrote
Hi @will-rowe, I'm done with my review.  All comments are in https://github.com/kuangmeng/CIMSA/issues, 1-19. Overall recommendation is "Reject, major revisions in order."
