### 2020-09-25 kuangmeng/CIMSA#3 @will-rowe
I would also like to see this but I should point out that JOSS don't explicitly require automated testing - it is just [preferred](https://joss.readthedocs.io/en/latest/review_criteria.html#tests).
