### 2020-09-25 kuangmeng/CIMSA#15 @will-rowe
Please advise here @kuangmeng. Did anyone else contribute directly to your submission? I read the paper acknowledgement as a generic thank you to those in the field - if this is correct, maybe rephrase it to avoid confusion (or remove if you want, as it's not required due to citations being sufficient)?
