### 2020-10-02 #2672 @xyzmjf
Hi Will. I have downloaded the code. I will be going over it this afternoon and this weekend. Updating the check list was a no go, as the invite has expired or been revoked. Mark
