### 2020-10-02 kuangmeng/CIMSA#1 @xyzmjf
I assumed that the actual folder extracted (CIMSA-master) should be used in place of the directory cimsa_folder mentioned in the readme.md file. The 'make clean' and 'make' commands seem to execute with no issues (but no feedback either). As above I found the 'node start_display.js' produced no output and simply seems to hang. Clearly the testing of the outcome an inexperienced user is likely to see from this code is far from adequate.

Running the executable 'cimsa' with the command ./cimsa would be expected to produce a usage example and input specification for this executable, and for this to be consistent with the example in the readme.md file. Instead the command ./cimsa leads to this output
Usage:
       mlpalign_main -G infile [> outfile] -- Get features for Classification on Posterior Probability Matrix.
       mlpalign_main -p <0|1|2> infile [> outfile] -- Use different Posterior Probability Matrix to do MSA.

This appears to refer to a different executable mplalign_main and the flags are not consistent with the readme.md desciption as far as I can tell.
