### 2020-10-06 kuangmeng/CIMSA#2 @xyzmjf
There is no documentaton on how to use the code mentioned in the manuscript. The expectation is that any potential user will start looking at the code and incorporate it themselves, which is possible but seems less than very likely. Again this is not encouraging the target audience.
