### 2020-10-06 kuangmeng/CIMSA#18 @xyzmjf
See comments on test script here:
https://github.com/kuangmeng/CIMSA/issues/20
There is no option to test only a user desired set of parameters (s1=0, s2=2,...etc.)
There is no guidance for the end user to make best use of the code.
