### 2020-10-06 kuangmeng/CIMSA#16 @xyzmjf
The manuscript the likely use case, that researchers into multiple sequence alignment might benefit from making use of this code. However, it seems only to cover the specific case of progressive pairwise alignments, and not simultaneous optimisation of the alignment(s) of multiple sequences. In any case I would expect the likely take up by these researchers to be very low. I suspect there is little real need for the availability of this  code in the field.
