### 2020-10-06 kuangmeng/CIMSA#15 @xyzmjf
There remains ambiguity regarding the authorship. If each of the routines is a redeveloped version of a previously published algorithm, then it would help greatly to mention the literature reference for these in the code. This would allow the correctness of implementation to be examined more easily. One would hope this is not just a copy paste of code from elsewhere, I assume it is not.
