### 2020-10-06 kuangmeng/CIMSA#16 @xyzmjf
In the statement of need in the manuscript the author states

> Till now, there is no efficient and accurate MSA tool of progressive strategy using Python
> programming language or providing a Python interface.

However the Biopython library clearly provides access of progressive MSA tools (such as ClustalW) with a Python interface.
https://biopython.org/wiki/AlignIO
Hence the assertion in the manuscript seems to be just plain wrong.
