### 2020-10-06 kuangmeng/CIMSA#11 @xyzmjf
The use of English needs to be much improved. The manuscript should first have been checked by someone with a longer history of English language academic publishing, before being submitted for publication. To be acceptable the manuscript would need careful rewriting.
