### 2020-10-06 kuangmeng/CIMSA#10 @xyzmjf
It is hard or impossible to link the code submitted to the publications/algorithms that are the first publication, or the publication used for re-implementation here. As a short manuscript there some mention of the progressive MSA algorithms, but no real description of pros and cons, hence no clear vision for the state of the field. The explanations given of each specific algorithm are unclear and not sufficient for the potential reader, who may not be an expert in the field, to be able to decipher the relative merits.

There are assertions like this

> What’s more, the partition function can
> also get very good results in this task.

With no real explanation as to what the partition function means.
