### 2020-10-06 #2672 @xyzmjf
Hi Will. I have seen enough now to come to a conclusion. The code and manuscript should be rejected in the current form. A lot of work on both would be needed to meet the standard expected for a peer reviewed publication. all of my comments can be found in the open issues. My main overall comments are here. https://github.com/kuangmeng/CIMSA/issues/19
