### 2020-10-08 #2672 @will-rowe
Dear @kuangmeng,

Unfortunately, after very thorough and fair reviews from @xyzmjf and @CFGrote, I don't think we can accept your paper for publication at JOSS at this point in time.

Although I can see the utility of the package, in its current form it does not meet our requirements for publication. As stated by the reviews, documentation is severely lacking and veracity of software cannot be established.

I think the ability to benchmark and parameterise MSA algorithms is a very appealing and important research goal; although it  is an often under-appreciated and under-performed task. I do think the constructive reviews can be used to bring this submission up to a standard acceptable to JOSS. In particular, I think the paper and README need to document a clear need and use case for this work. With this in mind, I think we can consider this a reject with major revisions needed. If you address these comments and concerns, pease feel free to tag me in a new submission and we can re-assess.

Thank you for your submission to JOSS and thank you again to the reviewers.
